requirements.txt
instagrapi>=2.1.5
tweepy>=4.14.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
pillow>=10.0.0
python-dateutil>=2.8.2
